added team route and outsourced the return samples
