+++ routes/returnSamples.js
@@ -1,0 +1,298 @@
+const returnSamples = {
+  player: {
+    meta: {
+      id: "1",
+      first_name: "Luka",
+      last_name: "Perkovic",
+      nationality: "Croatia",
+      ign: "Perkz",
+      team: "G2 Esports",
+      role: "mid",
+      tournaments: [
+        {
+          tournament_name: "LEC 2019 Spring",
+          placing: "1st",
+        },
+        {
+          tournament_name: "LEC 2019 Summer",
+          placing: "1st",
+        }
+      ],
+      avatar: "url",
+      rating: 96
+    },
+    "win-rate": {
+      win_rate: 56.775,
+      regional_rank: 3,
+      global_rank: 19,
+      is_above_average_rank: true,
+      global_average_win_rate: 43
+    },
+    kda: {
+      kda: 2.2,
+      regional_rank: 9,
+      global_rank: 54,
+      is_above_average_rank: true,
+      global_average_kda: 3.2
+    },
+    "kill-participation": {
+      kill_participation: 96,
+      regional_rank: 12,
+      global_rank: 33,
+      is_above_average_rank: true,
+      global_average_kill_participation: 56.3
+    },
+    "kill-share": {
+      kill_share: 35.4,
+      regional_rank: 9,
+      global_rank: 54,
+      is_above_average_rank: false,
+      global_average_kill_share: 27
+    },
+    "cs-per-minute": {
+      cs_per_minute: 11.6,
+      regional_rank: 12,
+      global_rank: 33,
+      is_above_average_rank: true,
+      global_average_cs_per_minute: 9.8
+    },
+    "unique-champions": {
+      unique_champions: 56,
+      regional_rank: 1,
+      global_rank: 1,
+      is_above_average_rank: true,
+      global_average_unique_champions: 25
+    },
+    "favorite-champions": [
+      {
+        champion: "Ryze",
+        image: "ryze_img",
+        win_rate: 55.8,
+        games_played: 43,
+        kda: 3.41
+      },
+      {
+        champion: "Orianna",
+        image: "orianna_img",
+        win_rate: 56.7,
+        games_played: 30,
+        kda: 4.18
+      },
+      {
+        champion: "Azir",
+        image: "azir_img",
+        win_rate: 72.4,
+        games_played: 29,
+        kda: 3.73
+      },
+      {
+        champion: "LeBlanc",
+        image: "leblanc_img",
+        win_rate: 71.4,
+        games_played: 28,
+        kda: 5.5
+      },
+      {
+        champion: "Syndra",
+        image: "syndra_img",
+        win_rate: 60,
+        games_played: 25,
+        kda: 3
+      },
+    ]
+  },
+  team: {
+    meta: {
+      id: "1",
+      team_name: "FlyQuest",
+      region: "LCS",
+      headquarters: "United States of America",
+      tournaments: [
+        {
+          tournament_name: "LEC 2019 Spring",
+          placing: "1st",
+        },
+        {
+          tournament_name: "LEC 2019 Summer",
+          placing: "1st",
+        }
+      ],
+      avatar: "url",
+      rating: 96
+    },
+    "win-rate": {
+      win_rate: 56.775,
+      total_games_played: 81,
+    },
+    kda: {
+      kda: 3.87,
+      regional_rank: 9,
+      global_rank: 54,
+      is_above_average_rank: false,
+      global_average_kda: 4.2
+    },
+    "favorite-champions": [
+      {
+        champion: "Ryze",
+        image: "ryze_img",
+        win_rate: 55.8,
+        games_played: 43,
+        kda: 3.41
+      },
+      {
+        champion: "Orianna",
+        image: "orianna_img",
+        win_rate: 56.7,
+        games_played: 30,
+        kda: 4.18
+      },
+      {
+        champion: "Azir",
+        image: "azir_img",
+        win_rate: 72.4,
+        games_played: 29,
+        kda: 3.73
+      },
+      {
+        champion: "LeBlanc",
+        image: "leblanc_img",
+        win_rate: 71.4,
+        games_played: 28,
+        kda: 5.5
+      },
+      {
+        champion: "Syndra",
+        image: "syndra_img",
+        win_rate: 60,
+        games_played: 25,
+        kda: 3
+      },
+    ],
+    "current-roster": [
+      {
+        avatar: "url",
+        role: "top",
+        id: 1,
+        ign: "Licorice",
+        first_name: "Eric",
+        last_name: "Ritchie",
+        nationality: "United States of America",
+        player_rating: 96,
+      },
+      {
+        avatar: "url",
+        role: "jungle",
+        id: 1,
+        ign: "Josedeodo",
+        first_name: "Brandon",
+        last_name: "Joel Villegas",
+        nationality: "Argentinian",
+        player_rating: 96,
+      },
+      {
+        avatar: "url",
+        role: "mid",
+        id: 1,
+        ign: "Palafox",
+        first_name: "Cristian",
+        last_name: "Palafox",
+        nationality: "United States of America",
+        player_rating: 96,
+      },
+      {
+        avatar: "url",
+        role: "bot",
+        id: 1,
+        ign: "Johnson",
+        first_name: "Johnson",
+        last_name: "Nguyen",
+        nationality: "United States of America",
+        player_rating: 96,
+      },
+      {
+        avatar: "url",
+        role: "support",
+        id: 1,
+        ign: "Diamond",
+        first_name: "David",
+        last_name: "Berube",
+        nationality: "United States of America",
+        player_rating: 96,
+      },
+      {
+        avatar: "url",
+        role: "sub support",
+        id: 1,
+        ign: "Diamond",
+        first_name: "David",
+        last_name: "Berube",
+        nationality: "United States of America",
+        player_rating: 96,
+      }
+    ],
+    "former-roster": [
+      {
+        avatar: "url",
+        role: "top",
+        id: 1,
+        ign: "Licorice",
+        first_name: "Eric",
+        last_name: "Ritchie",
+        nationality: "United States of America",
+        player_rating: 96,
+      },
+      {
+        avatar: "url",
+        role: "jungle",
+        id: 1,
+        ign: "Josedeodo",
+        first_name: "Brandon",
+        last_name: "Joel Villegas",
+        nationality: "Argentinian",
+        player_rating: 96,
+      },
+      {
+        avatar: "url",
+        role: "mid",
+        id: 1,
+        ign: "Palafox",
+        first_name: "Cristian",
+        last_name: "Palafox",
+        nationality: "United States of America",
+        player_rating: 96,
+      },
+      {
+        avatar: "url",
+        role: "bot",
+        id: 1,
+        ign: "Johnson",
+        first_name: "Johnson",
+        last_name: "Nguyen",
+        nationality: "United States of America",
+        player_rating: 96,
+      },
+      {
+        avatar: "url",
+        role: "support",
+        id: 1,
+        ign: "Diamond",
+        first_name: "David",
+        last_name: "Berube",
+        nationality: "United States of America",
+        player_rating: 96,
+      },
+      {
+        avatar: "url",
+        role: "sub support",
+        id: 1,
+        ign: "Diamond",
+        first_name: "David",
+        last_name: "Berube",
+        nationality: "United States of America",
+        player_rating: 96,
+      }
+    ]
+  }
+}
+
+module.exports = returnSamples;
