+++ routes/team/index.js
@@ -4,14 +4,14 @@ import verifyCardType from "../../utils/verifyCardType";
 
 const router = express.Router();
 
-router.get("/player/:player_id", async (req, res, next) => {
+router.get("/team/:team_id", async (req, res, next) => {
   const query = req.query;
-  const player_id = req.params.player_id;
+  const team_id = req.params.team_id;
 
-  if (!player_id) {
+  if (!team_id) {
     res.status(500).send({
       success: "false",
-      message: "Player not found"
+      message: "Team not found"
     })
     return;
   }
@@ -22,14 +22,14 @@ router.get("/player/:player_id", async (req, res, next) => {
     })
     return;
   }
-  if (!verifyCardType("player", query.card)) {
+  if (!verifyCardType("team", query.card)) {
     res.status(500).send({
       success: "false",
       message: "Invalid card name"
     })
     return;
   }
-  // if (player_id !== "exists") {
+  // if (team_id !== "exists") {
   //   res.status(404).send({
   //     success: "false",
   //     message: "Player not found"
@@ -38,7 +38,7 @@ router.get("/player/:player_id", async (req, res, next) => {
   // }
 
   // Handle each card function in a separate function, imported separately.
-  const data = await returnSamples.player[query.card];
+  const data = await returnSamples.team[query.card];
   res.status(200).send({
     success: 'true',
     data
